Create Request object
This is designed to take overrides, as the defaults selected may
not be wanted for every request. It's also designed to support
multiple function calls simultaneously: any number of function
objects (coming soon) can be loaded into a request, which is really
nice for Intacct, as often times several actions should be taken
simulatenously.