Add API Gateway
This one's a biggie. It takes the function-bundling request object,
applies the appropriate headers, and sends it off to Intacct.

Testing here is done largely through the use of spies and
mocha-style expectations, which explains some of the funny,
seemingly-unused arguments accepted by Api's #initialize and

Checked in a production environment with real calls, so I'm sure
it's 👍🏼