Add CreateProject function
Soon comes the UpdateProject function which, I see in advance,
will require reuse of some of the code that's specific for
ProjectCreation. With that in mind, I designed this modularly,
taking CreateCustomer and UpdateCustomer as an example.

One more thing to note here: I added a helper method to BaseFunction
to allow for the addition of custom fields to any function object.
Something tells me this isn't the last time we'll want some.