Change :id to :customerid in customer-related calls
This seems small, but it's important. I want the developer
interface for this gem to be as similar to the Intacct API as
possible. In Intacct calls that deal with customers, the ID is
called 'customerid,' so that's what it should be called here, too.

In making this change, I had to tweak the way that
ContactsHelper#contactname and #contact_params work. They'd
previously been looking for an ID field named :id, which will not
be consistent across different sets of attributes. Now the ID
needs to be provided explicitly.