Fix failing tests, improve coverage
Integration testing against VCR cassettes in another application
showed that the last commit introduced a bug: the new customerid
param wasn't being picked up for functions' controlids. This
fixes that problem and adds test coverage to ensure that it doesn't
happen again.

In a later commit, I'm going to do some refactoring on this spec.