Allows for dymanic function generation using DTD3.0 syntax.
This is a big deal. Long gone are the days of rigidly-ordered
XML. And with those days behind us, we can dynamically generate
any function that Intacct accepts.

The new syntax is as follows:

  request = Request.new(AUTH_PARAMS)
  request.create :customer, name: 'Some Name'
  request.update :project, recordno: '1234', name: 'New Name'
  request.send

It's as easy as that. Currently, this process supports all of the
function types defined in Functions::ALLOWED_TYPES. More can be
added later, if necessary.

Welcome to the future.