Allow for varying cases in request keys
The Intacct API calls for different format and cases in read and write requests.

For example, a read request would look like:

  request.readByQuery(nil, { object: 'VENDOR',
                             query: '',
                             fields: '*',
                             pagesize: 100})

Note: Object-Type VENDOR is passed as a value in key-value pair and the keys are lowercase.

A write request, on the other hand looks like:

  request.create(:CUSTOMER, { CUSTOMERID: '1',
                              FIRST_NAME: 'Han',
                              LAST_NAME: 'Solo',
                              TYPE: 'Person',
                              EMAIL1: 'han@solo.com',
                              STATUS: 'active'
                            })

Note: Object-Type CUSTOMER is passed as the first argument and is in uppercase along with the keys.

Before this commit, the gem was artificially upcasing all keys it was passed as hashes.
This meant that it did not work with any of the read queries. This commit removes
this iterative upcase instead keying on the case of the key which it is passed.It also adds
a new function type readMore to allowed function types.

This is a breaking change for create and update calls