Adds TravisCI and tests current Ruby versions
This commit ensures that TravisCI runs in an open manner with
each build that is pushed to privateprep/intacct_ruby

This includes a simple .travis.yml file that runs spec in all
currently maintained versions of ruby. I also added the travis
gem as a dev dependency and made sure these dependencies are
in ABC order.

You may notice that support is removed for Ruby 1.9.3. I chose
to remove support because the specified version of pry-byebug
requires ruby v2+. I think this gem would otherwise work with 1.9.3
but don't think we need to ensure function moving forward.
In this spirit, I also remove the .ruby-version file as testing
across several versions is now done in CI

This commit also restyles and adds a TravisCI badge to the README